Re-named all classes to be more appropriate and adjusted filenames accordingly
