+++ tests/fields/test-number-field.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Tests for the email field type.
+ * Tests for the number field type.
  *
  * @since 1.1.0
  *
@@ -10,31 +10,31 @@
 
 namespace HMCMB\Tests;
 
-use CMB_Email_Field;
+use CMB_Number_Field;
 
 /**
- * Class EmailFieldTestCase
+ * Class TestNumberField
  *
  * @group fields
  */
-class EmailFieldTestCase extends TestFieldCase {
+class TestNumberField extends TestFieldCase {
 	public function setUp() {
 		parent::setUp();
 
-		$this->instance = new CMB_Email_Field( 'CMB_Email_Field', 'Field', [] );
+		$this->instance = new CMB_Number_Field( 'CMB_Number_Field', 'Field', [] );
 	}
 
 	/**
 	 * Test that the number field outputs correctly against more specific field output.
 	 */
 	function testFieldOutput() {
-		$field        = new CMB_Email_Field( 'foo', 'Foo', array( 'hm@hmn.md' ) );
+		$field = new CMB_Number_Field( 'foo', 'Foo', array( 0.5 ), array( 'min' => 0.4, 'max' => 1 ) );
 
 		if ( ! self::$post ) {
 			$this->markTestSkipped( 'Post not found' );
 		}
 
-		$this->expectOutputRegex( '/(type=\"email\".*?id=\"foo-cmb-field-0\".*?value=\"hm@hmn.md\")/s' );
+		$this->expectOutputRegex( '/min="0.4".*max="1".*(type=\"number\".*?id=\"foo-cmb-field-0\".*?value=\"0.5\")/s' );
 
 		// Trigger output.
 		$field->html();
@@ -47,9 +47,9 @@ class EmailFieldTestCase extends TestFieldCase {
 	 */
 	public function valuesProvider() {
 		return [
-			[ [ 'A string' ] ],
-			[ [ 'hm@md.com' ] ],
-			[ [ 'mike@mike.co.uk' ] ],
+			[ [ '1' ] ],
+			[ [ 162735 ] ],
+			[ [ 0.5 ] ],
 		];
 	}
 }
